feature_engineering/cloud_functions/no_reference_qoe_http/requirements.txt
google-cloud-datastore 
opencv-contrib-python==4.1.1.26
numpy==1.17.3
pandas==0.24.2